Data Saved Differently
I discovered the shelve module, which I'm going to use to save the data
in a format that's much easier to use with Python. I'm hoping this will
help me get some more interesting usage stats for the upcoming season
16. This means I'm going to stop saving all of the text files from here
on, since the data will be accessible more easily if I leave it all in
one file.